// src/types/permission.ts
export enum Permission {
  VIEW_TICKETS = "VIEW_TICKETS",
  CREATE_TICKETS = "CREATE_TICKETS",
  EDIT_TICKETS = "EDIT_TICKETS",
  EDIT_TICKET_TITLE_AND_PRIORITY = "EDIT_TICKET_TITLE_AND_PRIORITY",
  DELETE_TICKETS = "DELETE_TICKETS",
  ASSIGN_TICKETS = "ASSIGN_TICKETS",

  VIEW_ANALYTICS = "VIEW_ANALYTICS",
  MANAGE_STAFF = "MANAGE_STAFF",
  MANAGE_ORGANIZATIONS = "MANAGE_ORGANIZATIONS",
}